tools/requirements-nsfw.txt
autokeras==3.0.0
tensorflow==2.21.0
torch==2.10.0
torchvision==0.25.0
open_clip_torch==3.3.0
pillow==12.1.1
numpy==2.4.3
opencv-python-headless==4.13.0.92
